% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkistra Algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{is a datafram}

\item{init}{is the initial node}
}
\value{
Shortest path from Initial Node
}
\description{
An algorithm to find the shortest path from a inital node to all other nodes.
Takes two inputs two find the shortest path from the initial node
}

