% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidian}
\alias{euclidian}
\title{Euclidean Algorithm}
\usage{
euclidian(a, b)
}
\arguments{
\item{a}{is numeric or integer of length 1}

\item{b}{is numeric or integer of length 1}
}
\value{
The greatest commom divisor betwen a and b
}
\description{
An algorithm to find the greatest common divisor of two numbers.
}
\references{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}

