% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuloh.R
\docType{package}
\name{kuloh}
\alias{kuloh}
\alias{kuloh-package}
\title{A package created for the lab 3 in the course Advanced Programming in R, LiU}
\description{
A package created for the lab 3 in the course Advanced Programming in R, LiU
}

