% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{A Dataset for the wiki_graph}
\format{An object of class \code{data.frame} with 18 rows and 3 columns.}
\source{
http://en.wikipedia.org/wiki/Dijkstra%27s_algorithm
}
\usage{
wiki_graph
}
\description{
The wiki_graph is a dataset that has 18 observation and 3 variables.
The name of the variables are "v1","v2" and "w". These repesent the nodes
and the weight of each node.
}
\details{
The wiki_graph can be used with the dijkstra function.
}
\keyword{datasets}

